% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{E_G_i_x_G_i_minus_1}
\alias{E_G_i_x_G_i_minus_1}
\title{Expected value of G_i(G_i - 1).}
\usage{
E_G_i_x_G_i_minus_1(i, alpha, C_mat)
}
\arguments{
\item{i}{Index of the node for which the expected value of G_i(G_i - 1) is computed.}

\item{alpha}{probability in \eqn{(0,1/2)} that an edge differs from that of the representative graph C.}

\item{C_mat}{A N \times N matrix encoding the adjacency matrix associated to the representative graph of the CER model.}
}
\value{
[numeric] Expected value of G_i(G_i - 1).
}
\description{
Computes the expected value of \eqn{G_i(G_i-1)} for node \code{i}
under the centered Erdős–Rényi model.
}
