% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expected_count_non_oriented_path2_CER}
\alias{expected_count_non_oriented_path2_CER}
\title{Expected number of non-oriented paths of length two.}
\usage{
expected_count_non_oriented_path2_CER(alpha, C_mat)
}
\arguments{
\item{alpha}{probability in \eqn{(0,1/2)} that an edge differs from that of the representative graph C.}

\item{C_mat}{A N \times N matrix encoding the adjacency matrix associated to the representative graph of the CER model.}
}
\value{
[numeric] Expected number of non-oriented paths of length two.
}
\description{
Computes the expected number of non-oriented paths of length two
under the centered Erdős–Rényi model.
}
