% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expected_count_triangles_CER}
\alias{expected_count_triangles_CER}
\title{Expected total number of triangles under a CER model.}
\usage{
expected_count_triangles_CER(alpha, C_mat)
}
\arguments{
\item{alpha}{[numeric] probability in \eqn{(0,1/2)} that an edge differs from that of the representative graph C.}

\item{C_mat}{[matrix] A N \times N matrix encoding the adjacency matrix associated to the representative graph of the CER model.}
}
\value{
[numeric] Expected total number of triangles in the graph.
}
\description{
Computes the expected total number of triangles in a random graph generated
from a centered Erdős--Rényi (CER) model with representative graph
\eqn{C}. The expectation depends on the Erdős--Rényi contribution,
the degree sequence of the representative graph, and the number of triangles
in the representative graph.
}
\details{
The expected triangle count is computed as
\deqn{
E[T(G)] =
\alpha^3 {N \choose 3}
+ \alpha^2(1-2\alpha)(N-2)d_C
+ \frac{1}{2}\alpha(1-2\alpha)^2
\left(\sum_i d_i^2 - 2d_C\right)
+ (1-2\alpha)^3 T(C)
}
where \eqn{d_i} denotes the degree of node \eqn{i},
\eqn{d_C=\sum_i d_i/2} is the number of edges in the reference graph,
and \eqn{T(C)} is the total number of triangles in \eqn{C}.
}
