% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expected_transitivity_CER_approx}
\alias{expected_transitivity_CER_approx}
\title{Approximate expected transitivity under the CER model.}
\usage{
expected_transitivity_CER_approx(alpha, C_mat)
}
\arguments{
\item{alpha}{probability in \eqn{(0,1/2)} that an edge differs from that of the representative graph C.}

\item{C_mat}{Reference graph defining the centered Erdős--Rényi model.}
}
\value{
[numeric] Approximate expected transitivity under the CER model.
}
\description{
Computes the approximation of the expected transitivity (global clustering
coefficient) under the centered Erdős–Rényi model.
}
