% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{G_from_CER}
\alias{G_from_CER}
\title{generate a CER-distributed graph.}
\usage{
G_from_CER(C, alpha)
}
\arguments{
\item{C}{[numeric] M-dimensional numeric binary vector encoding the lower triangular part of the adjacency matrix associated to the representative graph C.}

\item{alpha}{[numeric] probability an edge differs from that of the representative graph C.}
}
\value{
[numeric] M-dimensional numeric binary vector encoding the lower triangular part of the adjacency matrix associated to generated CER-distributed graph.
}
\description{
generate a CER-distributed graph.
}
