% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{G_from_GCER}
\alias{G_from_GCER}
\title{Generate a graph from a Generalized Centered Erdős–Rényi (GCER) model.}
\usage{
G_from_GCER(W, alpha)
}
\arguments{
\item{W}{M-dimensional numeric vector encoding the location parameter W.}

\item{alpha}{concentration parameter in \eqn{(0,1/2)}.}
}
\value{
[numeric] A binary M-dimensional vector encoding the lower triangular part of the generated graph.
}
\description{
Generate a graph from a Generalized Centered Erdős–Rényi (GCER) model.
}
