% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gCER_inference}
\alias{gCER_inference}
\title{Generalized CER model inference}
\usage{
gCER_inference(
  G_bar,
  n,
  prior_list = list(W = 1/2, n_0 = 1, a = 1, b = 1),
  mcmc = FALSE,
  mcmc_options = list(n_iter = 1000, burnin = 0, thinning = 1),
  parallel = TRUE
)
}
\arguments{
\item{G_bar}{[numeric vector] M-dimensional vector containing the edge-wise
sample means of the observed graphs, where M = N(N - 1)/2. The entries encode
the lower triangular part of the edge-by-edge average of the adjacency matrices
of the observed graphs of $G^{(1:n)}$,
\eqn{\bar{G} = n^{-1}\sum_{i=1}^n G_i}, assuming the observed graphs follow
a Centered Erdős–Rényi (CER) distribution.}

\item{n}{[integer] The sample size.}

\item{prior_list}{[list] Ordered list of prior hyperparameters:
the location parameter \code{W} and \eqn{n_0 > 0} of the GCER prior for
\eqn{C}; and the shape parameters \eqn{a,b > 0} of the truncated Beta prior
for the scale parameter \eqn{\alpha}.
\code{W} is an M-dimensional numeric vector with entries in \eqn{[0,1]}
encoding the lower triangular part of the representative graph. The default
value is \eqn{1/2} for every edge, reflecting prior ignorance and yielding a
semiconjugate model. By default, \eqn{n_0 = a = b = 1}.}

\item{mcmc}{[logical] If \code{TRUE}, posterior inference is performed via
MCMC. If \code{FALSE}, an empirical Bayes inferential approach is used.}

\item{mcmc_options}{[list] List of MCMC options, including
\code{n_iter}, \code{burnin}, and \code{thinning}.}

\item{parallel}{[logical] If \code{TRUE}, parallel computation is used from
the EB method to maximize the marginal likelihood w.r.t. alpha.}
}
\value{
[list] A list containing the inferential output.
}
\description{
Generalized CER model inference
}
