% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_sbm}
\alias{generate_sbm}
\title{generate a SBM-distributed graph.}
\usage{
generate_sbm(N, K = 2, intra_prob = 0.8, inter_prob = 0.1)
}
\arguments{
\item{N}{[numeric] number of nodes.}

\item{K}{[numeric] Number of blocks.}

\item{intra_prob}{[numeric] Probability of edges within blocks.}

\item{inter_prob}{[numeric] Probability of edges between blocks.}
}
\value{
[igraph] An igraph object representing a stochastic block model graph.
}
\description{
generate a SBM-distributed graph.
}
