% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_sbm_fast}
\alias{generate_sbm_fast}
\title{generate a SBM-distributed graph.}
\usage{
generate_sbm_fast(
  N,
  K = 2,
  intra_prob = 0.8,
  inter_prob = 0.1,
  z = sample(seq_len(K), N, TRUE)
)
}
\arguments{
\item{N}{[numeric] number of nodes.}

\item{K}{[numeric] Number of blocks.}

\item{intra_prob}{[numeric] Probability of edges within blocks.}

\item{inter_prob}{[numeric] Probability of edges between blocks.}
}
\value{
[matrix] N \times N binary matrix encoding the adjacency matrix associated to the generated SBM-distributed graph.
}
\description{
generate a SBM-distributed graph.
}
