% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rtbeta}
\alias{rtbeta}
\title{Generate random samples from a truncated Beta distribution.}
\usage{
rtbeta(n, alpha, beta, a = 0, b = 1)
}
\arguments{
\item{n}{[integer] Number of random samples to generate.}

\item{alpha}{[numeric] First shape parameter of the TBeta distribution.
Must be positive.}

\item{beta}{[numeric] Second shape parameter of the TBeta distribution.
Must be positive.}

\item{a}{[numeric] Lower truncation bound. Default is \eqn{0}.}

\item{b}{[numeric] Upper truncation bound. Default is \eqn{1/2}.}
}
\value{
[numeric] A vector of length \code{n} containing independent samples
from the truncated Beta distribution on the interval \eqn{(a,b)}.
}
\description{
This function generates random draws from a Beta distribution with shape
parameters \eqn{\alpha} and \eqn{\beta}, truncated to the interval
\eqn{(a,b)}.
}
