% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tau_alpha}
\alias{tau_alpha}
\title{Compute the tau transformation of alpha.
\deqn{\tau(\alpha,n_0) = \frac{\alpha^{n_0}}
{(1-\alpha)^{n_0}+\alpha^{n_0}}}
which maps a probability parameter \eqn{\alpha} into the interval
\eqn{(0,1/2)}.}
\usage{
tau_alpha(alpha, n_0)
}
\arguments{
\item{alpha}{[numeric] A probability parameter in the interval \eqn{(0,1/2)}.}

\item{n_0}{[numeric] A positive exponent controlling the concentration
of the transformation and interpreted as prior sample size.}
}
\value{
[numeric] The value of \eqn{\tau(\alpha,n_0)}.
}
\description{
Compute the tau transformation of alpha.
\deqn{\tau(\alpha,n_0) = \frac{\alpha^{n_0}}
{(1-\alpha)^{n_0}+\alpha^{n_0}}}
which maps a probability parameter \eqn{\alpha} into the interval
\eqn{(0,1/2)}.
}
